+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * fix dropdown closed by document click 
 * add Number.prototype.format
 * fix Metro Icon Font foe using with winphone apps
+* upd dialog, add runtime setContent* methods
 
 ## 3.0.14
 * fix tabcontrol frames overflow